Refactor. Add mounting from os env var
